Simplify docker/Dockerfile.ocaml
First, we make all of the steps into a single command.  This reduces
the amount of space that we use as there is now but a single layer.
We also take the opportunity to reduce repetition by using an env
var.  And we wrap long lines while we are here.