test_runtime.sh: stop comparing expected text to previous text
We remove the code which looks for the previous prompt.  As it turns
out: when you reduce the polling delay, we have seen that not only
can you obtain a partially written new prompt---but you can also get a
partially erased old prompt. Given that we expect the tests to succeed,
we do not gain much from short-circuiting the test early and so if we
just compare what we expect until it times out that is likely sufficient
for now. This may change, if we want to write tests that fail because
we expect the wrong text, but we do not think that is likely.