s/strncpy/strlcpy/
strncpy doesn't necessarily nul-terminate strings.  We also use
sizeof(arg0) so that you can visually validate that we are measuring
the correct buffers without xrefing the header files.

A small oddness in here: apparently, BOLOS' strncpy(3) zeroes the
entire buffer and when we switched to strlcpy(3) we needed to zero
few chars here and there or we'd occasionally output some old state.