app: fix dropped 'right button' presses
TLDR; sometimes when a user presses right, we start parsing again and
find that we need more data. We then put ourselves into a state of
requesting more data.

That data then comes, we parse it, and once we have a screen we push
it. Then we redisplay the screen.

*BUT* the screen we redisplay is not the next screen, but rather the
current one. Since `tz_ui_stream()` goes straight back to the top
level, in this _one_ instance, we never go to the next screen, until
the user presses right again.

Therefore, we check in `tz_ui_stream()` whether the user pressed right
previously and, if so, we display the next screen.