integration tests: only run debugging app if the normal app fails.
We do this because the normal app is much quicker than the debugging
app.  Considering that the normal case should be that each test
succeeds, we should make the normal case much faster with this
change.

We store both failures unconditionally in integration_tests.json,
but we do only count it as a single failure.