keys.c: simplify READ_UNALIGNED_BIG_ENDIAN()
We were reversing the bytes into a temp buffer.  The code is much simpler,
easier to read, and more portable if we use the resulting typed int
directly shifting the bytes.

We also reduce the size of the name by s/BIG_ENDIAN/BE/.

We could just use ntohl(3), honestly, as this macro is only used for 1
byte quantities where it does nothing and 4 byte quantities.