Operation parser: do not make small break between inside long data
When printing binary data, if the data was too long for the 'capture'
buffer, the data was printed and the buffer could then continue to
capture the data.
However, the printing would display all the data captured for the
moment, and the buffer would then move on to the next screen to
display the rest, which would split the printing of the data.
This is no longer the case.