Fix a couple of issues with tz_string_to_mutez()
scan-build turned up an issue which is that *res could be NULL in
the error case, because we jump to it after testing it to be NULL.
This commit eliminates the goto error logic and just returns false,
we do this because there are no cleanup actions to take.  We don't
bother zeroing out res on error because the caller should not rely
on this behaviour: if the function returns false, the result is
undefined.

We also eliminate using memset(3) to zero a uint64_t.  It is easier and
easier to read to just use normal assignment (``='').

We also eliminate a variable ``r'' which isn't being used.